-rewrote to use array and client class -organized the js code into js folder
