+++ index.html
@@ -54,7 +54,7 @@
         </tr>
     </table>
 
-    <script src="script.js"></script>
+    <script type="module" src="js/form.js"></script>
 </body>
 </html>
