+++ js/client.js
@@ -1,0 +1,8 @@
+export class Client {
+    constructor(firstName, surname, age, mobileNumber ){
+        this.firstName = firstName
+        this.surname = surname
+        this.age = age
+        this.mobileNumber = mobileNumber
+    }
+}
